e2fsck: use floating point number to calculate the fragmented ratio
If there are many fragmented files such that 10000 * ctx->fs_fragmented
exceeds the maximal number of __u32, e2fsck might show an incorrect
value for a percent of non-contiguous files due to an overflow. The
ratio of non-contiguous directories and total inodes also have the same
issue.

This commit fixes this by using floating point arithmetic. It has enough
precision to hold the percent of fragmented files/dirs/total up to the
second decimal place.

Signed-off-by: Daichi Mukai <daichi@mukai.dev>